ci: Run GH workflows on public workers
The self-hosted runner cluster "Rudibert" will be decomissioned for
internal reasons. Therefore, the workflows need to run on public
workers. It is possible that, due to that, the molecule job may fail
for external contributors because the TTS token will be set via an
environment variable in the project settings and not via the worker
settings anymore.

Signed-off-by: Tom Siewert <tom.siewert@hetzner.com>